split hello_world example
