+++ crates/ui/Cargo.toml
@@ -15,6 +15,37 @@ doctest = false
 decimal = ["dep:rust_decimal"]
 inspector = []
 webview = ["dep:wry"]
+# For syntax highlighting in Markdown and CodeEditor.
+tree-sitter-languages = [
+    "dep:tree-sitter-bash",
+    "dep:tree-sitter-c",
+    "dep:tree-sitter-c-sharp",
+    "dep:tree-sitter-cmake",
+    "dep:tree-sitter-cpp",
+    "dep:tree-sitter-css",
+    "dep:tree-sitter-diff",
+    "dep:tree-sitter-elixir",
+    "dep:tree-sitter-embedded-template",
+    "dep:tree-sitter-go",
+    "dep:tree-sitter-graphql",
+    "dep:tree-sitter-html",
+    "dep:tree-sitter-java",
+    "dep:tree-sitter-javascript",
+    "dep:tree-sitter-jsdoc",
+    "dep:tree-sitter-make",
+    "dep:tree-sitter-md",
+    "dep:tree-sitter-proto",
+    "dep:tree-sitter-python",
+    "dep:tree-sitter-ruby",
+    "dep:tree-sitter-rust",
+    "dep:tree-sitter-scala",
+    "dep:tree-sitter-sequel",
+    "dep:tree-sitter-swift",
+    "dep:tree-sitter-toml-ng",
+    "dep:tree-sitter-typescript",
+    "dep:tree-sitter-yaml",
+    "dep:tree-sitter-zig"
+]
 
 [dependencies]
 gpui.workspace = true
@@ -61,35 +92,35 @@ chrono = "0.4.38"
 # Code Editor
 indexset = "0.12.2"
 tree-sitter = "0.25.4"
-tree-sitter-bash = "0.23.3"
-tree-sitter-c = "0.24.1"
-tree-sitter-c-sharp = "0.23.1"
-tree-sitter-cmake = "0.7.1"
-tree-sitter-cpp = "0.23.4"
-tree-sitter-css = "0.23.2"
-tree-sitter-diff = "0.1.0"
-tree-sitter-elixir = "0.3"
-tree-sitter-embedded-template = "0.23.0"
-tree-sitter-go = "0.23.4"
-tree-sitter-graphql = "0.1.0"
-tree-sitter-html = "0.23.2"
-tree-sitter-java = "0.23.5"
-tree-sitter-javascript = "0.23.1"
-tree-sitter-jsdoc = "0.23.2"
 tree-sitter-json = "0.24.8"
-tree-sitter-make = "1.1.1"
-tree-sitter-md = { git = "https://github.com/tree-sitter-grammars/tree-sitter-markdown", tag = "v0.5.0" }
-tree-sitter-proto = "0.2.0"
-tree-sitter-python = "0.23.6"
-tree-sitter-ruby = "0.23.1"
-tree-sitter-rust = "0.24.0"
-tree-sitter-scala = "0.23.4"
-tree-sitter-sequel = "0.3.8"
-tree-sitter-swift = "0.7.0"
-tree-sitter-toml-ng = "0.7.0"
-tree-sitter-typescript = "0.23.2"
-tree-sitter-yaml = "0.7.1"
-tree-sitter-zig = "1.1.2"
+tree-sitter-bash = { version = "0.23.3", optional = true }
+tree-sitter-c = { version = "0.24.1", optional = true }
+tree-sitter-c-sharp = { version = "0.23.1", optional = true }
+tree-sitter-cmake = { version = "0.7.1", optional = true }
+tree-sitter-cpp = { version = "0.23.4", optional = true }
+tree-sitter-css = { version = "0.23.2", optional = true }
+tree-sitter-diff = { version = "0.1.0", optional = true }
+tree-sitter-elixir = { version = "0.3", optional = true }
+tree-sitter-embedded-template = { version = "0.23.0", optional = true }
+tree-sitter-go = { version = "0.23.4", optional = true }
+tree-sitter-graphql = { version = "0.1.0", optional = true }
+tree-sitter-html = { version = "0.23.2", optional = true }
+tree-sitter-java = { version = "0.23.5", optional = true }
+tree-sitter-javascript = { version = "0.23.1", optional = true }
+tree-sitter-jsdoc = { version = "0.23.2", optional = true }
+tree-sitter-make = { version = "1.1.1", optional = true }
+tree-sitter-md = { git = "https://github.com/tree-sitter-grammars/tree-sitter-markdown", tag = "v0.5.0", optional = true }
+tree-sitter-proto = { version = "0.2.0", optional = true }
+tree-sitter-python = { version = "0.23.6", optional = true }
+tree-sitter-ruby = { version = "0.23.1", optional = true }
+tree-sitter-rust = { version = "0.24.0", optional = true }
+tree-sitter-scala = { version = "0.23.4", optional = true }
+tree-sitter-sequel = { version = "0.3.8", optional = true }
+tree-sitter-swift = { version = "0.7.0", optional = true }
+tree-sitter-toml-ng = { version = "0.7.0", optional = true }
+tree-sitter-typescript = { version = "0.23.2", optional = true }
+tree-sitter-yaml = { version = "0.7.1", optional = true }
+tree-sitter-zig = { version = "1.1.2", optional = true }
 
 [dev-dependencies]
 indoc = "2"
